Added podreview entity

+++ Models/CupSize.cs
@@ -3,6 +3,6 @@ namespace NespressoReviewsApi.Models
     public class CupSize : BaseModel
     {
         public string Name { get; set; }
-        public float Volume { get; set; }
+        public double Volume { get; set; }
     }
 }
